+++ src/Oro/Bundle/FilterBundle/Provider/DatagridFiltersProvider.php
@@ -4,8 +4,8 @@ namespace Oro\Bundle\FilterBundle\Provider;
 
 use Oro\Bundle\DataGridBundle\Datagrid\Common\DatagridConfiguration;
 use Oro\Bundle\DataGridBundle\Extension\Formatter\Configuration as FormatterConfiguration;
-use Oro\Bundle\DataGridBundle\Extension\Formatter\Property\PropertyInterface;
 use Oro\Bundle\FilterBundle\Factory\FilterFactory;
+use Oro\Bundle\FilterBundle\Filter\FilterUtility;
 use Oro\Bundle\FilterBundle\Grid\Extension\Configuration;
 
 /**
@@ -13,6 +13,9 @@ use Oro\Bundle\FilterBundle\Grid\Extension\Configuration;
  */
 class DatagridFiltersProvider implements DatagridFiltersProviderInterface
 {
+    public const ORDER_FIELD_NAME  = 'order';
+    public const ORDER_FIELD_LABEL = 'label';
+
     private FilterFactory $filterFactory;
 
     private string $applicableDatasourceType;
@@ -32,24 +35,76 @@ class DatagridFiltersProvider implements DatagridFiltersProviderInterface
             return [];
         }
 
-        $filtersConfig = $gridConfig->offsetGetByPath(Configuration::COLUMNS_PATH, []);
+        $filtersConfig = $this->getSortedFilters($gridConfig);
 
+        $filters = [];
         foreach ($filtersConfig as $filterName => $filterConfig) {
-            if (!empty($filterConfig[PropertyInterface::DISABLED_KEY])) {
-                // Skips disabled filter.
-                continue;
-            }
-
-            // If label is not set, tries to use corresponding column label.
-            if (!isset($filterConfig['label'])) {
-                $filterConfig['label'] = $gridConfig->offsetGetByPath(
-                    sprintf('[%s][%s][label]', FormatterConfiguration::COLUMNS_KEY, $filterName)
-                );
-            }
-
             $filters[$filterName] = $this->filterFactory->createFilter($filterName, $filterConfig);
         }
 
-        return $filters ?? [];
+        return $filters;
+    }
+
+    /**
+     * @SuppressWarnings(PHPMD.CyclomaticComplexity)
+     * @SuppressWarnings(PHPMD.NPathComplexity)
+     *
+     * @param DatagridConfiguration $gridConfig
+     * @return array
+     */
+    private function getSortedFilters(DatagridConfiguration $gridConfig): array
+    {
+        $filtersConfig = $gridConfig->offsetGetByPath(Configuration::COLUMNS_PATH, []);
+        foreach ($filtersConfig as $filterName => &$filterConfig) {
+            foreach ([self::ORDER_FIELD_LABEL, self::ORDER_FIELD_NAME, FilterUtility::DISABLED_KEY] as $field) {
+                if (!array_key_exists($field, $filterConfig)) {
+                    $filterConfig[$field] = $gridConfig->offsetGetByPath(
+                        sprintf('[%s][%s][%s]', FormatterConfiguration::COLUMNS_KEY, $filterName, $field)
+                    );
+                }
+            }
+
+            // BC layer start
+            if (isset($filterConfig['enabled'])) {
+                if (!$filterConfig['enabled']) {
+                    $filterConfig[FilterUtility::RENDERABLE_KEY] = false;
+                }
+                unset($filterConfig['enabled']);
+            }
+            // BC layer end
+
+            if (!empty($filterConfig[FilterUtility::DISABLED_KEY])) {
+                unset($filtersConfig[$filterName]);
+            }
+        }
+        unset($filterConfig);
+
+        $weight  = 1;
+        $defined = array_column($filtersConfig, self::ORDER_FIELD_NAME);
+        foreach ($filtersConfig as &$filterConfig) {
+            $order = filter_var(
+                $filterConfig[self::ORDER_FIELD_NAME] ?? null,
+                FILTER_VALIDATE_INT
+            );
+            if ($order === false) {
+                while (\in_array($weight, $defined, true)) {
+                    $weight++;
+                }
+                $defined[] = $weight;
+                $order = $weight;
+            }
+
+            $filterConfig[self::ORDER_FIELD_NAME] = $order;
+        }
+        unset($filterConfig);
+
+        uasort($filtersConfig, static function (array $a, array $b) {
+            if ($a[self::ORDER_FIELD_NAME] === $b[self::ORDER_FIELD_NAME]) {
+                return 0;
+            }
+            return ($a[self::ORDER_FIELD_NAME] < $b[self::ORDER_FIELD_NAME]) ? -1 : 1;
+        });
+
+        return $filtersConfig;
     }
 }